Refactor DI components, improve docs and organization
Move DI component types to separate files and add XML documentation for improved clarity. Enhance OpenFeatureBuilder and Guard with better comments and parameter docs. Document OpenFeatureOptions members and methods. Remove obsolete commented code. Improves code readability, maintainability, and API discoverability.
